+++ DAL/Entities/HopDong.cs
@@ -36,7 +36,7 @@ namespace DAL.Entities
         [StringLength(20)]
         public string PPTHANHTOAN { get; set; }
 
-        public decimal? TONGTT { get; set; }
+        public double? TONGTT { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<CT_HOPDONG> CT_HOPDONG { get; set; }
